Issue 149:  	 Fragment Minibrowser shows a little off from the position whrere it should be when it maximized.
Update issue 149
Status: Fixed
First, this problem is occured only when fixedPortalHader is true.

When height is not "auto", hasLayout property become true in IE6.
When hasLayout is true, an element which position is absolute is influenced by it's parent node for unexplained reasons.
So to fix this problem I set panel height to "auto" before maximizing a gadget and adjust panel height after closing a maximized gadget.

git-svn-id: https://infoscoop.googlecode.com/svn/trunk@279 d4e23b88-c782-11de-b6dc-25c59a28ecba