Issue 208:  The function to change font size improperly works when it should not work.
Update issue 208
Status: Fixed
The following code is for a case that a font size of body is changed by css.
 getActiveStyle( document.body, "font-size");
So this code cannot be deleted.
IS_Portal.defaultFontSize should be a font size of body set by css.
So I set a font size of body set by user after getting IS_Portal.defaultFontSize.

git-svn-id: https://infoscoop.googlecode.com/svn/trunk@341 d4e23b88-c782-11de-b6dc-25c59a28ecba