Issue 217:  	 When column width is resized, adjustHeight function is called with undefined parameter.In case of using Safari or Chrome, a height of gadget is set to wrong length.
Update issue 217
Status: Fixed
The description of this issue is wrong.
This is the following problem.
When adjust column, infoScoop calls adjustHeight() for all Gadgets which use dynamic-height feature.
There are the follwoing Gadgets.
* Gadget declares dynamic-height feature, but doesn't call adjustHeight.
* Gadget calls adjustHeight function with an argument. For example, adjustHeight(100).
In these cases, calling adjustHeight() causes unexpected behavior.

So I modified to call adjustHeight() only when a Gadget calls adjustHeight without arguments.

git-svn-id: https://infoscoop.googlecode.com/svn/trunk@475 d4e23b88-c782-11de-b6dc-25c59a28ecba